+++ R/monolixReadData.R
@@ -176,11 +176,27 @@ rxUiGet.monolixOmega <- function(x, ...) {
 }
 attr(rxUiGet.monolixOmega, "rstudio") <- lotri::lotri(a~0.1)
 
+#' Monolix name of an nlmixr2 population parameter
+#'
+#' @param name nlmixr2 theta name
+#' @param muRef mu-referenced theta to Monolix variable names
+#' @param covDataFrame mu-referenced covariate data frame
+#' @return `beta_<var>_<covariate>` for a mu-referenced covariate effect,
+#'   otherwise `<var>_pop`
+#' @noRd
+.monolixPopParName <- function(name, muRef, covDataFrame) {
+  .w <- which(covDataFrame$covariateParameter == name)
+  if (length(.w) == 1) {
+    return(paste0("beta_", muRef[covDataFrame$theta[.w]], "_",
+                  covDataFrame$covariate[.w]))
+  }
+  paste0(muRef[name], "_pop")
+}
+
 .monolixGetPopParValue <- function(name, muRefCurEval, muRef, covDataFrame, pop) {
   .w <- which(covDataFrame$covariateParameter == name)
   if (length(.w) == 1) {
-    .par <- paste0("beta_", muRef[covDataFrame$theta[.w]], "_",
-                   covDataFrame$covariate[.w])
+    .par <- .monolixPopParName(name, muRef, covDataFrame)
     .w <- which(pop$parameter == .par)
     if (length(.w) != 1) return(NA_real_)
     return(pop$value[.w])
@@ -346,6 +362,9 @@ rxUiGet.monolixEtaObf <- function(x, ...) {
   .muRef <- c(.split$pureMuRef, .split$taintMuRef)
   .etaMonolix <- rxUiGet.monolixIndividualParameters(x, ...)
   if (is.null(.etaMonolix)) return(NULL)
+  # the SAEM etas (posterior means from the last SAEM iterations), like
+  # nlmixr2est's own saem; the conditional mode (`_mode`) is the FOCEi
+  # family's kind of empirical Bayes estimate
   .n <- c("id", vapply(.etas$neta1, function(i) {
     paste0("eta_",   .mlxtranGetIndividualMuRefEtaMonolixName(.ui, i, .muRef), "_SAEM")
   }, character(1), USE.NAMES=FALSE))
@@ -455,10 +474,14 @@ rxUiGet.monolixCovariance <- function(x, ...) {
     .sa <- FALSE
   }
   .j <- rxUiGet.monolixJacobian(x, ...)
-  .n <- vapply(dimnames(.j)[[1]], function(n){
-    paste0(.muRef[n], "_pop")
-  }, character(1), USE.NAMES=FALSE)
+  # a covariate effect is beta_<var>_<covariate>, not <var>_pop
+  .covDataFrame <- .ui$saemMuRefCovariateDataFrame
+  .n <- vapply(dimnames(.j)[[1]], .monolixPopParName, character(1),
+               muRef=.muRef, covDataFrame=.covDataFrame, USE.NAMES=FALSE)
   .cov <- .cov[.n, .n]
+  # nlmixr2's parameter names whichever Monolix version wrote the matrix
+  # (only the pre-2020/2021 conversion below used to rename it)
+  dimnames(.cov) <- dimnames(.j)
   .ui <- x[[1]]
   rxode2::rxAssignControlValue(.ui, ".covMethod", ifelse(.sa, "MonolixSA", "MonolixLin"))
   if (.monolixCovarianceNeedsConversion(x, .sa)) {
@@ -549,12 +572,14 @@ rxUiGet.monolixPreds <- function(x, ...) {
     .tmp$CMT <- paste(.tmp$CMT)
   }
   .ret <- merge(fit$ui$monolixPreds, .tmp, by=.by)
+  # the same (SAEM) individual estimates the fit's etas came from
+  .ipred <- .ret$indivPred_SAEM
   .ci <- (1 - fit$monolixControl$ci) / 2
   .q <- c(0, .ci, 0.5, 1 - .ci, 1)
-  .qi <- stats::quantile(with(.ret, 100*abs((IPRED-indivPred_SAEM)/indivPred_SAEM)), .q, na.rm=TRUE)
+  .qi <- stats::quantile(100*abs((.ret$IPRED - .ipred)/.ipred), .q, na.rm=TRUE)
   .qp <- stats::quantile(with(.ret, 100*abs((PRED-popPred)/popPred)), .q, na.rm=TRUE)
-  .qai <- stats::quantile(with(.ret, abs(IPRED-indivPred_SAEM)), .q, na.rm=TRUE)
-  .qap <- stats::quantile(with(.ret, abs((PRED-popPred)/popPred)), .q, na.rm=TRUE)
+  .qai <- stats::quantile(abs(.ret$IPRED - .ipred), .q, na.rm=TRUE)
+  .qap <- stats::quantile(with(.ret, abs(PRED-popPred)), .q, na.rm=TRUE)
   .sigdig <- 3
   .msg <- c(paste0("IPRED relative difference compared to Monolix IPRED: ", round(.qi[3], 2),
                  "%; ", fit$monolixControl$ci * 100,"% percentile: (",
@@ -569,7 +594,7 @@ rxUiGet.monolixPreds <- function(x, ...) {
             paste0("PRED absolute difference compared to Monolix PRED: atol=",
                    signif(.qap[3], digits=.sigdig),
                    "; ", fit$monolixControl$ci * 100,"% percentile: (",
-                   signif(.qap[2], digits=.sigdig), ", ", signif(.qp[4], digits=.sigdig), ")"))
+                   signif(.qap[2], digits=.sigdig), ", ", signif(.qap[4], digits=.sigdig), ")"))
   list(individualRel=.qi , popRel=.qp,
        individualAbs=.qai, popAbs=.qap,
        message=.msg)